rework input/outputs API
