+++ examples/hls.rs
@@ -44,7 +44,7 @@ fn start_example_client_code() -> Result<()> {
             "transport_protocol": "tcp_server",
             "port": INPUT_PORT,
             "video": {
-                "codec": "h264"
+                "decoder": "ffmpeg_h264"
             }
         }),
     )?;
@@ -70,19 +70,24 @@ fn start_example_client_code() -> Result<()> {
                     "width": VIDEO_RESOLUTION.width,
                     "height": VIDEO_RESOLUTION.height,
                 },
-                "encoder_preset": "medium",
+                "encoder": {
+                    "type": "ffmepg_h264",
+                    "preset": "fast",
+                },
                 "initial": {
-                    "type": "shader",
-                    "id": "shader_node_1",
-                    "shader_id": "shader_example_1",
-                    "children": [
-                        {
-                            "id": "input_1",
-                            "type": "input_stream",
-                            "input_id": "input_1",
-                        }
-                    ],
-                    "resolution": { "width": VIDEO_RESOLUTION.width, "height": VIDEO_RESOLUTION.height },
+                    "root": {
+                        "type": "shader",
+                        "id": "shader_node_1",
+                        "shader_id": "shader_example_1",
+                        "children": [
+                            {
+                                "id": "input_1",
+                                "type": "input_stream",
+                                "input_id": "input_1",
+                            }
+                        ],
+                        "resolution": { "width": VIDEO_RESOLUTION.width, "height": VIDEO_RESOLUTION.height },
+                    }
                 }
             }
         }),
